wip restructuring users (#59)
* wip restructuring users

* start switching to user-specific directories instead of prefixes

+++ install.sh
@@ -8,12 +8,17 @@
 
 set -euo pipefail
 
+get_user() {
+  # user account name aliases should be defined here
+	echo "${USER/jbagdis/jeff}"
+}
+
 init_colors() {
   set +u
   # Use colors, but only if connected to a terminal
-	# and only if that terminal supports them.
+  # and only if that terminal supports them.
   if command -v tput >/dev/null 2>&1; then
-      num_colors=$(tput colors)
+    num_colors=$(tput colors)
   fi
   if [ -t 1 ] && [ -n "$num_colors" ] && [ "$num_colors" -ge 8 ]; then
     RED="$(tput setaf 1)"
@@ -34,18 +39,18 @@ init_colors() {
 }
 
 init_variables() {
-	# set default values
-	set +u
-	if [ -z "${ENV_GIT_DIR}" ]; then
-	  ENV_GIT_DIR=~/.env.git
-	fi
-	set -u
+  # set default values
+  set +u
+  if [ -z "${ENV_GIT_DIR}" ]; then
+    ENV_GIT_DIR=~/.env.git
+  fi
+  set -u
 }
 
 create_directories_if_needed() {
   # Ensure ~/.ssh/sockets directory exits
   mkdir -p ~/.ssh/sockets
-  
+
   # Ensure ipython configuration directory exists
   mkdir -p ~/.ipython/profile_default
 }
@@ -53,28 +58,27 @@ create_directories_if_needed() {
 create_or_update_links() {
   # (re-)link all files from the env git repo into the home directory
   printf "${BLUE}Linking environment components into home directory...\n${NORMAL}"
-  link_home_dir bin ""
-  link_dot_file gitconfig ""
-  link_dot_file gitconfig.user "_$(get_user)"
-  link_dot_file gitignore_global ""
-  link_dot_file inputrc ""
-  link_dot_file oh-my-zsh ""
-  link_dot_file profile ""
-  link_dot_file zprofile ""
-  link_dot_file zshrc ""
+  link_home_dir bin
+  link_dot_file gitconfig
+  link_dot_file gitconfig.user
+  link_dot_file gitignore_global
+  link_dot_file inputrc
+  link_dot_file oh-my-zsh
+  link_dot_file profile
+  link_dot_file zprofile
+  link_dot_file zshrc
   link_ssh_file authorized_keys "_$(get_user)"
-  link_ssh_file config ""
+  link_ssh_file config
   link_dot_file p10k.zsh "_$(get_user)"
-  link_dot_file "ipython/profile_default/ipython_config.py" ""
-  link_dot_file "tmux.conf" ""
+  link_dot_file "ipython/profile_default/ipython_config.py"
+  link_dot_file "tmux.conf"
 
   # shellcheck disable=SC2012
-  ls "${ENV_GIT_DIR}/LaunchAgents" | while read -r file
-  do
-    link_launch_agent "$file" ""
+  ls "${ENV_GIT_DIR}/LaunchAgents" | while read -r file; do
+    link_launch_agent "$file"
     launchctl load ~/Library/LaunchAgents/"$file" 2>&1 | while read -r line; do
-        printf "\t\t${YELLOW}> %s${NORMAL}\n" "$line"
-      done
+      printf "\t\t${YELLOW}> %s${NORMAL}\n" "$line"
+    done
   done
 }
 
@@ -97,49 +101,71 @@ set_shell_to_zsh() {
   fi
 }
 
-get_user() {
-	echo "${USER/jbagdis/jeff}"
-}
-
 abstract_link() {
+  # get the input arguments
+  # (the fourth argument (source suffix) is optional)
+  set +u
+  SRC_SUFFIX="$4"
+  if [ -z "${SRC_SUFFIX}" ]; then
+    SRC_SUFFIX=""
+  fi
+  set -u
   SRC_PREFIX="$1"
   DEST_PREFIX="$2"
   FILE="$3"
-  SRC_SUFFIX="$4"
   SRC="${SRC_PREFIX}${FILE}${SRC_SUFFIX}"
+  USER_SRC="users/$(get_user)/${SRC_PREFIX}${FILE}${SRC_SUFFIX}"
   DEST="${DEST_PREFIX}${FILE}"
-  if [ -L ~/"${DEST}" ]; then
-    # file is a link; don't bother backing up
-    rm ~/"${DEST}"
-    true
-  else
-    if [ -e ~/"${DEST}" ]; then
-      printf "\tBacking up ${DEST}\n"
-      if [ -d ~/"${DEST}.bak" ]; then
-        rm -rf ~/"${DEST}.bak"
-      fi
-      mv ~/"${DEST}" ~/"${DEST}.bak"
-    fi
+  # use the user-specific source if it exists
+  echo "${ENV_GIT_DIR}/${USER_SRC}"
+  if [ -e "${ENV_GIT_DIR}/${USER_SRC}" ]; then
+    SRC="${USER_SRC}"
   fi
-  printf "${GREEN}\tLinking ${DEST}\n${NORMAL}"
-  ln -sf "${ENV_GIT_DIR}/${SRC}" ~/"${DEST}"
-  #echo "ln -sf \"${ENV_GIT_DIR}/${SRC}\" ~/\"${DEST}\""
+  # ensure that the source exists before trying to link
+  if [ ! -e "${ENV_GIT_DIR}/${SRC}" ]; then
+    printf "\t${YELLOW}Not linking '${DEST}' because '${SRC}' does not exist.${NORMAL}\n"
+    touch ~/"${DEST}"
+  else
+    if [ -L ~/"${DEST}" ]; then
+      # file is a link; don't bother backing up
+      rm ~/"${DEST}"
+      true
+    else
+      if [ -e ~/"${DEST}" ]; then
+        printf "\tBacking up ${DEST}\n"
+        if [ -d ~/"${DEST}.bak" ]; then
+          rm -rf ~/"${DEST}.bak"
+        fi
+        mv ~/"${DEST}" ~/"${DEST}.bak"
+      fi
+    fi
+    printf "${GREEN}\tLinking ${DEST}\n${NORMAL}"
+    ln -sf "${ENV_GIT_DIR}/${SRC}" ~/"${DEST}"
+  fi
 }
 
 link_dot_file() {
+  set +u
   abstract_link "dot_files/" "." "$1" "$2"
+  set -u
 }
 
 link_home_dir() {
+  set +u
   abstract_link "home_dirs/" "" "$1" "$2"
+  set -u
 }
 
 link_ssh_file() {
+  set +u
   abstract_link "ssh/" ".ssh/" "$1" "$2"
+  set -u
 }
 
 link_launch_agent() {
+  set +u
   abstract_link "LaunchAgents/" "Library/LaunchAgents/" "$1" "$2"
+  set -u
 }
 
 install_check_prereqs() {
@@ -177,13 +203,13 @@ update_check_existing_env() {
 }
 
 install_env_git_do_clone() {
-  env git clone --progress https://github.com/jbagdis/env.git "${ENV_GIT_DIR}" && \
-  pushd "${ENV_GIT_DIR}" && \
-  git remote set-url --push origin git@github.com:jbagdis/env.git && \
-  pushd dot_files && \
-  env git clone --progress https://github.com/ohmyzsh/ohmyzsh.git oh-my-zsh && \
-  popd && \
-  popd
+  env git clone --progress https://github.com/jbagdis/env.git "${ENV_GIT_DIR}" &&
+    pushd "${ENV_GIT_DIR}" &&
+    git remote set-url --push origin git@github.com:jbagdis/env.git &&
+    pushd dot_files &&
+    env git clone --progress https://github.com/ohmyzsh/ohmyzsh.git oh-my-zsh &&
+    popd &&
+    popd
 }
 
 install_env_git() {
@@ -196,11 +222,11 @@ install_env_git() {
   #  as Windows ACLs take precedence over umasks
   #  except for filesystems mounted with option "noacl".
   umask g-w,o-w
-  
+
   printf "${BLUE}Cloning Environment Repository...\n${NORMAL}"
   # The Windows (MSYS) Git is not compatible with normal use on cygwin
   if [ "${OSTYPE}" = cygwin ]; then
-    if git --version | grep msysgit > /dev/null; then
+    if git --version | grep msysgit >/dev/null; then
       printf "${RED}\tError: Windows/MSYS Git is not supported on Cygwin${NORMAL}"
       printf "\tMake sure the Cygwin git package is installed and is first on the path"
       exit 1
@@ -215,14 +241,14 @@ install_env_git() {
 }
 
 update_env_git() {
-    # update the env git repo
-    pushd "${ENV_GIT_DIR}"
-    if [ "$(git rev-parse --abbrev-ref HEAD)" = "main" ]; then
-      git pull
-    else
-      printf "${YELLOW}\tYou are not on the 'main' branch; skipping 'git pull'.\n${NORMAL}"
-    fi
-    popd
+  # update the env git repo
+  pushd "${ENV_GIT_DIR}"
+  if [ "$(git rev-parse --abbrev-ref HEAD)" = "main" ]; then
+    git pull
+  else
+    printf "${YELLOW}\tYou are not on the 'main' branch; skipping 'git pull'.\n${NORMAL}"
+  fi
+  popd
 }
 
 install_or_update_powerlevel_10k() {
@@ -257,7 +283,7 @@ install() {
   create_or_update_links
   install_or_update_powerlevel_10k
   set_shell_to_zsh
-  printf "${BOLD}Shell Environment successfully installed.\n${NORMAL}" 
+  printf "${BOLD}Shell Environment successfully installed.\n${NORMAL}"
   printf "${BLUE}Memoizing new profile...\n${NORMAL}"
   env zsh -l
 }
